Create main.js
What did you like about this project?
I did not like anything.
What did you struggle with in this project?
i think the project is too overwhelming for me as the topic of node.js has just been taught and i still have not grasp it well enough.  I did not know what to expect and do not know how to proceed with self learning. Maybe some more video guides or examples of how to use it would be good. I do not know how to even search for the method or answers to proceed with my difficulties.
What would make your experience with this assessment better?
I think more simpler exercises could be given out as practice. My understanding of npm is still really vague and i cannot really process how it gels together.  I think maybe we should be given the resources to learn before the project would have been good. As i think without a guide and by aimlessly learning on it is quite inefficient especially if we read up on irrelevant topics which will not be beneficial for our assessment.